Extract es2k-specific public functions from ovsp4rt.cc
